feat: refactor blog post schema and update import path for PageContainer component
