+++ types/index.ts
@@ -1,2 +1,1 @@
 export * from './entities';
-export * from './elements';
